+++ e2e/tests/plugins.e2e.js
@@ -37,8 +37,6 @@ describe('plugins', () => {
     after(() => cy.eyesClose());
 
     it('render image toolbar and settings', function() {
-      cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE).shrinkPlugin();
-      cy.eyesCheckWindow(this.test.title + '  - plugin toolbar');
       cy.openImageSettings();
       cy.get(`[data-hook=${IMAGE_SETTINGS.PREVIEW}]:first`);
       cy.eyesCheckWindow(this.test.title + ' - settings');
@@ -50,6 +48,14 @@ describe('plugins', () => {
       cy.eyesCheckWindow(this.test.title + ' - delete image title');
       cy.openImageSettings(false).addImageLink();
       cy.eyesCheckWindow(this.test.title + ' - add a link');
+      cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE).pluginSizeOriginal();
+      cy.eyesCheckWindow(this.test.title + '  - plugin original size');
+      cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE).shrinkPlugin();
+      cy.eyesCheckWindow(this.test.title + '  - plugin toolbar');
+      cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE).pluginSizeBestFit();
+      cy.eyesCheckWindow(this.test.title + '  - plugin content size');
+      cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE).pluginSizeFullWidth();
+      cy.eyesCheckWindow(this.test.title + '  - plugin full width size');
     });
   });