+++ packages/editor-common/web/src/index.js
@@ -82,6 +82,7 @@ export {
   getFocusedBlockKey,
   calculateDiff,
   getPostContentSummary,
+  createSelection,
 } from './Utils/draftUtils';
 export { isiOS } from './Utils/isiOS';
 export { getSelectionStyles } from './Utils/inlineStyleUtils';