+++ CHANGELOG.md
@@ -15,6 +15,11 @@
 </details>
 <hr/>
 
+## 6.10.4 (Mar 15, 2020)
+### :bug: Bug Fix
+- `editor`
+  - [d486af87](https://github.com/wix-incubator/rich-content/commit/d486af87) fix plain text paste
+
 ## 6.10.2 (Mar 15, 2020)
 ### :house: Internal
 - `wrapper`